% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainfunctions.R
\name{fit_ordinal}
\alias{fit_ordinal}
\title{Fit the cumulative logistic model}
\usage{
fit_ordinal(ttnsr,C,A_1,A_2,A_3,omega=TRUE,alpha = TRUE)
}
\arguments{
\item{ttnsr}{an observed tensor data}

\item{C}{an initial point for a core tensor}

\item{A_1, A_2, A_3}{initial points for factor matrices}

\item{omega}{cut-off points if it is known,

\code{omega = TRUE} if it is unknown}

\item{alpha}{a signal level

\code{alpha = TRUE} if there is no constraint for max-norm of latent tensor parameters}
}
\value{
a list containing the following:

\code{C} - an estimated core tensor

\code{A_1, A_2, A_3} - estimated factor matrices

\code{theta} - an estimated latent parameter tensor

\code{iteration} - the number of iterations

\code{cost} - evaluations of cost functions at each iterations

\code{omega} - estimated cut-off points vector (k-level)
}
\description{
Fitting the cumulative logistic model with alternating optimization method.
}
\examples{
# draw random ordinal tensor
indices = c(10,10,10)
rank = c(3,3,3)
otensor <- array(sample(c(1,2,3),prod(indices),replace = TRUE),dim = indices)

#Initial points for `fit_ordinal' function
A_1 = pracma::randortho(indices[1])[,1:rank[1]]
A_2 = pracma::randortho(indices[2])[,1:rank[2]]
A_3 = pracma::randortho(indices[3])[,1:rank[3]]
C = rTensor::rand_tensor(modes = rank)*10

#Estimation for parameters
parameter_est = fit_ordinal(otensor,C,A_1,A_2,A_3,omega = TRUE,alpha = 100)
}
\references{
Lee and Wang. “Tensor denoising and completion based on ordinal observations.” (2020).
}
